Fixed issue in Counterparty credit limit Update. Used float instead of int in process table. [Issue: 38204]
git-svn-id: http://dev.farrms.us:18080/pioneer/products/trmtracker/branches/TRMTracker_Enercity/Ver_4.2.003@113333 fd52c265-6bb1-e54c-ad91-b501145a01df

+++ Database/StoredProc/spa_counterparty_credit_limits.sql
@@ -236,7 +236,7 @@ BEGIN
 	                 		counterparty_credit_limit_id INT,
 	                 		counterparty_name  VARCHAR(200),
 	                 		debt_rating        VARCHAR(200),
-	                 		credit_limit       INT,
+	                 		credit_limit       float,
 	                 		as_of_date			VARCHAR(30),
 	                 		hyperlink1 VARCHAR(5000), 
          					hyperlink2 VARCHAR(5000),
